Use https in Dockerfile
Install temp CA and SSL cert so that the SSL port can be exposed.
We use a flexible and valid SSL cert from Azure in the hosting environment; this cert is only used to support https all the way through so that the underlying HttpContext can generate valid https links for us.